Split DiskGraphReader away from (now abstract) GraphReader interface
