+++ src/mjolnir/valhalla_convert_transit.cc
@@ -21,7 +21,7 @@
 #include "baldr/datetime.h"
 #include "baldr/graphconstants.h"
 #include "baldr/graphid.h"
-#include "baldr/graphreader.h"
+#include "baldr/diskgraphreader.h"
 #include "baldr/graphtile.h"
 #include "baldr/tilehierarchy.h"
 #include "filesystem.h"
@@ -974,7 +974,7 @@ void build_tiles(const boost::property_tree::ptree& pt,
   stats.dep_count = 0;
   stats.midnight_dep_count = 0;
 
-  GraphReader reader_transit_level(pt);
+  DiskGraphReader reader_transit_level(pt);
   auto database = pt.get_optional<std::string>("timezone");
   // Initialize the tz DB (if it exists)
   sqlite3* tz_db_handle = GetDBHandle(*database);
@@ -1011,7 +1011,7 @@ void build_tiles(const boost::property_tree::ptree& pt,
 
     GraphId transit_tile_id = GraphId(tile_id.tileid(), tile_id.level() + 1, tile_id.id());
     const GraphTile* transit_tile = reader_transit_level.GetGraphTile(transit_tile_id);
-    GraphTileBuilder tilebuilder_transit(reader_transit_level.tile_dir(), transit_tile_id, false);
+    GraphTileBuilder tilebuilder_transit(reader_transit_level.GetTileDir(), transit_tile_id, false);
 
     auto tz = DateTime::get_tz_db().from_index(DateTime::get_tz_db().to_index("America/New_York"));
     uint32_t tile_creation_date =
