+++ valhalla/baldr/connectivity_map.h
@@ -3,6 +3,8 @@
 
 #include <valhalla/baldr/pathlocation.h>
 
+#include <boost/property_tree/ptree.hpp>
+
 #include <cstdint>
 #include <unordered_map>
 #include <unordered_set>
